+++ persim/landscapes/auxiliary.py
