+++ persim/landscapes/base.py
