+++ persim/landscapes/exact.py
@@ -5,9 +5,9 @@
 import itertools
 import numpy as np
 from operator import itemgetter
-from .pers_landscape_aux import union_crit_pairs
-from .pers_landscape import PersLandscape
-from .pers_landscape_approx import PersLandscapeApprox
+from .auxiliary import union_crit_pairs
+from .base import PersLandscape
+from .approximate import PersLandscapeApprox
 
 __all__ = ["PersLandscapeExact", "vectorize"]
