+++ persim/landscapes/visuals.py
@@ -8,9 +8,9 @@ import matplotlib as mpl
 from matplotlib import cm
 import matplotlib.pyplot as plt
 import numpy as np
-from .pers_landscape import PersLandscape
-from .pers_landscape_exact import PersLandscapeExact
-from .pers_landscape_approx import PersLandscapeApprox
+from .base import PersLandscape
+from .exact import PersLandscapeExact
+from .approximate import PersLandscapeApprox
 
 mpl.rcParams["text.usetex"] = True
